+++ test/sql/engines.sql
@@ -79,6 +79,8 @@ SELECT a, percentile_cont(0.75) WITHIN GROUP (ORDER BY f) / sum(d) FROM t4 GROUP
 EXPLAIN (VERBOSE, COSTS OFF) SELECT a, percentile_disc(0.75) WITHIN GROUP (ORDER BY f) FROM t4 GROUP BY a;
 EXPLAIN (VERBOSE, COSTS OFF) SELECT a, percentile_disc(0.75) WITHIN GROUP (ORDER BY f) / sum(d) FROM t4 GROUP BY a;
 
+EXPLAIN (VERBOSE, COSTS OFF) SELECT a, count(e) FROM t4 GROUP BY a;
+
 DROP USER MAPPING FOR CURRENT_USER SERVER engines_loopback;
 CALL clickhouse_perform('engines_admin', 'DROP DATABASE engines_test');
 DROP SERVER engines_loopback CASCADE;